Clear every cross-bucket edge out of agent-surface
34 value-level edges left agent-surface for creation, delivery, catalog and
community. Each one is answered by the cheapest honest mechanism, in this
order: import platform directly, relocate genuinely shared code, reclassify a
module that only agent-surface ever reads, extract a mixed concern, and only
then inject.

Redirected, no code moved: InvalidUploadError and the upload caps were reached
through games-store.ts's re-exports of platform/upload-error.ts and
platform/games-repo-contract.ts; deriveGateStatusString through gate-verdict.ts's
re-export of @gamedevpl/contract; PROPOSAL_NO_JOB through proposals.ts's
re-export of platform/proposal-limits.ts. mcp-server.ts's GamesStore import used
the mixed `import { type X }` form the rule does not exempt, and is now a
whole-statement `import type`.

Relocated to platform, because nothing in the owning bucket's vocabulary is
closed over: gateCrashStall, isActiveBuildRound and allowsSelfToPlatformHandoff
are pure JobState/JobStall predicates and join job-state.ts (builder.ts
re-exports the two it used to own, for creation's own callers); parseGameMedia
and IN_FLIGHT_RUN_STATUSES are a self-contained JSON parser and a GitHub run
vocabulary that catalog only happened to host; the transcript window caps are
numbers the MCP get_transcript schema advertises; source-patch.ts and
delivery-path-guard.ts move whole, having no domain import at all beyond the
`diff` package and the delivery file list; agent-state.ts is the vendor-neutral
state vocabulary the store, job-state and three buckets already read type-only.

Reclassified and physically moved into agent-surface, because each has zero
consumers inside the bucket it lived in and exists for the agent surface alone
-- the rule the map already applies to agent-build-brief.ts and
game-manifest-hint.ts: example-files.ts, agent-tasks.ts, build-prompt.ts (its
only import is agent-surface's own BuildBrief type), and gate-verdict.ts, whose
own header says it is "shared by the channel's gate route and MCP `start`" and
whose sole delivery dependency is a type-only GamesStore.

Extracted rather than moved or injected: agent-backend-env.ts mixed its own
agent-backend registry with creation's game-seeding env wiring -- eight of the
34 edges, none of whose results agent-surface consumes. That half is now
creation/seed-provider-env.ts and app.ts calls it directly, so nothing needs
injecting into agent-surface at all.

Injected at the composition root, because the code is real domain logic that
belongs where it is: computeStageAdvisories (already carrying creation's
typecheck preflight as injected params, now bound once in app.ts and handed to
the channel), loadBuildTranscript (same shape as the isPresenceEventText seam it
already had), assertDeliverableSourcePath (closes over delivery's own
RESERVED_SEGMENTS/ALLOWED_SOURCE_FILES -- relocating it would move the delivery
contract away from the store that enforces it), community's proposal state
machine, and catalog's GitHub client factory for copilot's Actions-run cancel.

SourceDeliveryValidationError is deliberately not collapsed into the base
InvalidUploadError the way creator-code.ts's checks were in #1149:
agent-channel.ts reads error.reason, which only the subclass carries, so
delivery exports an isSourceDeliveryValidationError narrowing predicate and
app.ts injects it. self-build.test.ts covers both reason values.

Baselines follow the code: a whole-file move carries its comment-prose and
module-size keys to the new path, the agent-backend-env split reseals both
halves at 141 + 123 against a 264 baseline (zero net growth), and job-state.ts
rises by exactly what builder.ts and gate-crash.ts shed.
