+++ apps/api/src/agent-surface/agent-backend-env.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { randomBytes, randomUUID } from 'node:crypto';
-import { createAgentBackendRegistryFromEnv, createSeedProvidersFromEnv } from './agent-backend-env.js';
+import { createAgentBackendRegistryFromEnv } from './agent-backend-env.js';
 import { registerManagedProvider, type ManagedProviderConfig } from './managed-agent.js';
 import { createGeminiManagedProvider } from './managed-provider-gemini.js';
 
@@ -27,6 +27,17 @@ const ENV_KEYS = [
 ] as const;
 
 const MCP_URL = 'https://www.gamedev.pl/api/mcp';
+
+// N1: app.ts supplies catalog's client; the registry builds none.
+const GITHUB_STUB = {
+  deleteBranch: async () => {},
+  listWorkflowRuns: async () => [],
+  cancelWorkflowRun: async () => {},
+};
+
+function registryFromEnv(log: { info: () => void; warn: () => void }) {
+  return createAgentBackendRegistryFromEnv(log, undefined, { githubClientFactory: () => GITHUB_STUB });
+}
 
 describe('createAgentBackendRegistryFromEnv', () => {
   const previous = new Map<string, string | undefined>();
@@ -60,7 +71,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_COPILOT_MCP_REPO: 'gamedevpl/scratchpad',
     });
     const warn = vi.fn();
-    const registry = createAgentBackendRegistryFromEnv({ info: vi.fn(), warn });
+    const registry = registryFromEnv({ info: vi.fn(), warn });
 
     // The default vendor's own backend failed to build...
     expect(registry.platformByVendor.has('anthropic')).toBe(false);
@@ -83,7 +94,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_MCP_URL: MCP_URL,
       MANAGED_AGENT_COPILOT_MCP_REPO: 'gamedevpl/scratchpad',
     });
-    const registry = createAgentBackendRegistryFromEnv({ info: vi.fn(), warn: vi.fn() });
+    const registry = registryFromEnv({ info: vi.fn(), warn: vi.fn() });
     expect(registry.defaultVendor).toBeUndefined();
     expect(registry.platformByVendor.has('copilot')).toBe(true);
     expect(registry.self.name).toBe('self');
@@ -101,7 +112,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_MCP_URL: MCP_URL,
       AGENT_TASKS_TOKEN: randomBytes(32).toString('hex'),
     });
-    const registry = createAgentBackendRegistryFromEnv({ info: vi.fn(), warn: vi.fn() });
+    const registry = registryFromEnv({ info: vi.fn(), warn: vi.fn() });
     expect(registry.platformByVendor.get('anthropic')?.name).toBe('managed:anthropic');
     expect(registry.platformByVendor.has('gemini')).toBe(false);
   });
@@ -125,7 +136,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
         GEMINI_API_KEY: `gemini-${randomUUID()}`,
         AGENT_TASKS_TOKEN: randomBytes(32).toString('hex'),
       });
-      const registry = createAgentBackendRegistryFromEnv({ info: vi.fn(), warn: vi.fn() });
+      const registry = registryFromEnv({ info: vi.fn(), warn: vi.fn() });
       expect(registry.platformByVendor.has('gemini')).toBe(true);
       expect(seen?.agentId).toBeUndefined();
       expect(seen?.environmentId).toBeUndefined();
@@ -146,7 +157,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_COPILOT_MCP_CUSTOM_AGENT: 'game-builder-mcp',
     });
     const info = vi.fn();
-    const registry = createAgentBackendRegistryFromEnv({ info, warn: vi.fn() });
+    const registry = registryFromEnv({ info, warn: vi.fn() });
 
     expect(registry.platformByVendor.get('copilot')?.name).toBe('managed:copilot');
     expect(info).toHaveBeenCalledWith(
@@ -164,7 +175,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_COPILOT_MCP_REPO: 'gamedevpl/scratchpad',
     });
     const warn = vi.fn();
-    const registry = createAgentBackendRegistryFromEnv({ info: vi.fn(), warn });
+    const registry = registryFromEnv({ info: vi.fn(), warn });
 
     expect(registry.platformByVendor.get('copilot')).toBeUndefined();
     expect(warn).toHaveBeenCalledWith(
@@ -182,7 +193,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_MAX_SECONDS: '900',
     });
     const warn = vi.fn();
-    const registry = createAgentBackendRegistryFromEnv({ info: vi.fn(), warn });
+    const registry = registryFromEnv({ info: vi.fn(), warn });
 
     expect(registry.platformByVendor.get('copilot')).toBeUndefined();
     expect(warn).toHaveBeenCalledWith(
@@ -202,7 +213,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       AGENT_TASKS_TOKEN: `copilot-${randomUUID()}`,
     });
     const info = vi.fn();
-    const registry = createAgentBackendRegistryFromEnv({ info, warn: vi.fn() });
+    const registry = registryFromEnv({ info, warn: vi.fn() });
 
     expect(registry.platformByVendor.get('gemini')?.name).toBe('managed:gemini');
     expect(info).toHaveBeenCalledWith(
@@ -223,7 +234,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       AGENT_TASKS_TOKEN: `copilot-${randomUUID()}`,
     });
     const info = vi.fn();
-    createAgentBackendRegistryFromEnv({ info, warn: vi.fn() });
+    registryFromEnv({ info, warn: vi.fn() });
 
     expect(info).toHaveBeenCalledWith(
       expect.objectContaining({ vendor: 'gemini', model: 'gemini-3.8-flash' }),
@@ -240,7 +251,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_MAX_TOTAL_TOKENS: '0',
     });
     const warn = vi.fn();
-    const registry = createAgentBackendRegistryFromEnv({ info: vi.fn(), warn });
+    const registry = registryFromEnv({ info: vi.fn(), warn });
 
     expect(registry.platformByVendor.size).toBe(0);
     expect(warn).toHaveBeenCalledWith(
@@ -256,7 +267,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_MAX_SECONDS: '900',
     });
     const warn = vi.fn();
-    const registry = createAgentBackendRegistryFromEnv({ info: vi.fn(), warn });
+    const registry = registryFromEnv({ info: vi.fn(), warn });
 
     expect(registry.platformByVendor.size).toBe(0);
     expect(warn).toHaveBeenCalledWith(
@@ -275,7 +286,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_MCP_URL: MCP_URL,
     });
     const info = vi.fn();
-    const registry = createAgentBackendRegistryFromEnv({ info, warn: vi.fn() });
+    const registry = registryFromEnv({ info, warn: vi.fn() });
 
     expect(registry.platformByVendor.get('openai')?.name).toBe('managed:openai');
     expect(info).toHaveBeenCalledWith(
@@ -292,7 +303,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_MCP_URL: MCP_URL,
     });
     const warn = vi.fn();
-    const registry = createAgentBackendRegistryFromEnv({ info: vi.fn(), warn });
+    const registry = registryFromEnv({ info: vi.fn(), warn });
 
     expect(registry.platformByVendor.has('openai')).toBe(false);
     expect(warn).toHaveBeenCalledWith(
@@ -310,7 +321,7 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_MAX_SECONDS: '900',
       MANAGED_AGENT_MCP_URL: MCP_URL,
     });
-    const registry = createAgentBackendRegistryFromEnv({ info: vi.fn(), warn: vi.fn() });
+    const registry = registryFromEnv({ info: vi.fn(), warn: vi.fn() });
 
     expect(registry.platformByVendor.has('anthropic')).toBe(false);
     expect(registry.platformByVendor.get('openai')?.name).toBe('managed:openai');
@@ -328,90 +339,12 @@ describe('createAgentBackendRegistryFromEnv', () => {
       MANAGED_AGENT_MAX_SECONDS: undefined,
     });
     const warn = vi.fn();
-    const registry = createAgentBackendRegistryFromEnv({ info: vi.fn(), warn });
+    const registry = registryFromEnv({ info: vi.fn(), warn });
 
     expect(registry.platform).toBeUndefined();
     expect(warn).toHaveBeenCalledWith(
       expect.objectContaining({ vendor }),
       expect.stringContaining('MANAGED_AGENT_MAX_SECONDS'),
-    );
-  });
-});
-
-const SEED_ENV_KEYS = [
-  'SEED_PROVIDER',
-  'SEED_MODEL',
-  'SEED_MAX_OUTPUT_TOKENS',
-  'SEED_PICK_MAX_OUTPUT_TOKENS',
-  'SEED_ANTHROPIC_API_KEY',
-  'SEED_ANTHROPIC_MODEL',
-  'SEED_ANTHROPIC_MAX_OUTPUT_TOKENS',
-  'SEED_ANTHROPIC_PICK_MAX_OUTPUT_TOKENS',
-  'SEED_META_API_KEY',
-  'SEED_META_MODEL',
-  'SEED_META_MAX_OUTPUT_TOKENS',
-  'SEED_META_PICK_MAX_OUTPUT_TOKENS',
-] as const;
-
-describe('createSeedProvidersFromEnv', () => {
-  const previous = new Map<string, string | undefined>();
-
-  afterEach(() => {
-    for (const key of SEED_ENV_KEYS) {
-      const value = previous.get(key);
-      if (value === undefined) delete process.env[key];
-      else process.env[key] = value;
-    }
-    previous.clear();
-  });
-
-  function setEnv(values: Partial<Record<(typeof SEED_ENV_KEYS)[number], string | undefined>>): void {
-    for (const key of SEED_ENV_KEYS) {
-      if (!previous.has(key)) previous.set(key, process.env[key]);
-      const next = values[key];
-      if (next === undefined) delete process.env[key];
-      else process.env[key] = next;
-    }
-  }
-
-  it('defaults Meta to an 8192 pick budget, unlike every other vendor', () => {
-    setEnv({ SEED_META_API_KEY: 'k', SEED_META_MODEL: 'muse-spark-1.2' });
-    const { providers } = createSeedProvidersFromEnv();
-    expect(providers.get('meta')?.pickMaxOutputTokens).toBe(8192);
-    expect(providers.get('vertex')?.pickMaxOutputTokens).toBeUndefined();
-  });
-
-  it('lets an explicit env var override the Meta default', () => {
-    setEnv({ SEED_META_API_KEY: 'k', SEED_META_MODEL: 'muse-spark-1.2', SEED_META_PICK_MAX_OUTPUT_TOKENS: '4096' });
-    const { providers } = createSeedProvidersFromEnv();
-    expect(providers.get('meta')?.pickMaxOutputTokens).toBe(4096);
-  });
-
-  it('reads a per-vendor pick budget for a vendor with no built-in default', () => {
-    setEnv({
-      SEED_ANTHROPIC_API_KEY: 'k',
-      SEED_ANTHROPIC_MODEL: 'claude-haiku-4-5',
-      SEED_ANTHROPIC_PICK_MAX_OUTPUT_TOKENS: '3000',
-    });
-    const { providers } = createSeedProvidersFromEnv();
-    expect(providers.get('anthropic')?.pickMaxOutputTokens).toBe(3000);
-  });
-
-  it('reads a Vertex-wide pick budget from the unprefixed var', () => {
-    setEnv({ SEED_PICK_MAX_OUTPUT_TOKENS: '5000' });
-    const { providers } = createSeedProvidersFromEnv();
-    expect(providers.get('vertex')?.pickMaxOutputTokens).toBe(5000);
-  });
-
-  it('ignores a non-positive-integer pick budget and warns', () => {
-    setEnv({ SEED_META_API_KEY: 'k', SEED_META_MODEL: 'muse-spark-1.2', SEED_META_PICK_MAX_OUTPUT_TOKENS: 'nope' });
-    const warn = vi.fn();
-    const { providers } = createSeedProvidersFromEnv({ info: vi.fn(), warn });
-    // Falls back to the Meta-specific default, not the raw garbage value.
-    expect(providers.get('meta')?.pickMaxOutputTokens).toBe(8192);
-    expect(warn).toHaveBeenCalledWith(
-      expect.objectContaining({ envVar: 'SEED_META_PICK_MAX_OUTPUT_TOKENS' }),
-      expect.stringContaining('not a positive integer'),
     );
   });
 });
