+++ apps/api/src/agent-surface/agent-channel-examples.ts
@@ -1,12 +1,7 @@
 import type { FastifyInstance, FastifyReply, FastifyRequest } from 'fastify';
 import { AGENT_CHANNEL_ROUTES } from '@gamedevpl/contract';
 import { getAgentBuildExample, listAgentBuildExamples } from './agent-build-examples.js';
-import {
-  ExampleFilesError,
-  listExampleFiles,
-  readExampleFile,
-  type ExampleFileStore,
-} from '../creation/example-files.js';
+import { ExampleFilesError, listExampleFiles, readExampleFile, type ExampleFileStore } from './example-files.js';
 import { exampleUnpackCommand } from '../platform/kit-registry.js';
 import { DEFAULT_SIGNED_URL_TTL_SECONDS, type GcsObjectStore } from '../delivery/gcs-sign.js';
 import type { AgentTokenAccess } from '../platform/agent-token.js';
