+++ apps/api/src/agent-surface/agent-channel-kit.ts
@@ -9,7 +9,7 @@ import {
   parseKitRegistry,
   parseKitSidecar,
 } from '../platform/kit-registry.js';
-import type { GateVerdictSummary } from '../delivery/gate-verdict.js';
+import type { GateVerdictSummary } from './gate-verdict.js';
 import type { AgentTokenAccess } from '../platform/agent-token.js';
 import type { Store, SubmissionRecord } from '../platform/store.js';
