+++ apps/api/src/agent-surface/agent-game-key-resolve.ts
@@ -16,7 +16,7 @@ import {
   type GameAgentKeyClaims,
 } from './agent-game-key.js';
 import { InvalidAgentTokenError } from '../platform/agent-token.js';
-import { isActiveBuildRound } from '../creation/builder.js';
+import { isActiveBuildRound } from '../creation/job-state.js';
 import type { GameAgentKeyRecord, Store, SubmissionRecord } from '../platform/store.js';
 import { creatorOwnsSlug } from '../platform/slug-ownership.js';
