+++ apps/api/src/agent-surface/managed-backend.ts
@@ -1,6 +1,6 @@
 // AgentBackend over any ManagedAgentProvider, dispatched over MCP.
 import type { AgentBackend, BuildBrief, DispatchResult, SeedDelivery } from './agent-backend.js';
-import { buildPrompt } from '../delivery/build-prompt.js';
+import { buildPrompt } from './build-prompt.js';
 import type { AgentObservation, AgentSessionTokens } from '../creation/job-state.js';
 import { appendKitDigest, type KitDigestLoader } from './kit-digest.js';
 import {
