+++ apps/api/src/agent-surface/managed-provider-copilot.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it, vi } from 'vitest';
 import { randomBytes } from 'node:crypto';
-import type { AgentTask, AgentTaskInput, AgentTasksClient } from '../creation/agent-tasks.js';
-import { buildPrompt } from '../delivery/build-prompt.js';
+import type { AgentTask, AgentTaskInput, AgentTasksClient } from './agent-tasks.js';
+import { buildPrompt } from './build-prompt.js';
 import type { BuildBrief } from './agent-backend.js';
 import {
   COPILOT_AGENT_WORKFLOW_PATH,
