+++ apps/api/src/platform/agent-state.ts
