+++ apps/api/src/platform/github-run-status.ts
@@ -1,0 +1,3 @@
+// Statuses meaning a GitHub Actions run still burns time.
+
+export const IN_FLIGHT_RUN_STATUSES = ['queued', 'in_progress', 'requested', 'waiting', 'pending'];
