+++ eslint-rules/comment-prose-baseline.json
@@ -17,7 +17,7 @@
     "apps/api/scripts/verify-erase-signals.ts": 217,
     "apps/api/src/agent-build-media.test.ts": 667,
     "apps/api/src/agent-build-reads.test.ts": 14,
-    "apps/api/src/agent-surface/agent-backend-env.ts": 264,
+    "apps/api/src/agent-surface/agent-backend-env.ts": 141,
     "apps/api/src/agent-surface/agent-backend.ts": 654,
     "apps/api/src/agent-surface/agent-build-brief.test.ts": 21,
     "apps/api/src/agent-surface/agent-build-brief.ts": 95,
@@ -31,8 +31,12 @@
     "apps/api/src/agent-surface/agent-game-key-resolve.ts": 302,
     "apps/api/src/agent-surface/agent-game-key.ts": 346,
     "apps/api/src/agent-surface/agent-session-revocation.ts": 81,
+    "apps/api/src/agent-surface/agent-tasks.test.ts": 197,
+    "apps/api/src/agent-surface/agent-tasks.ts": 1135,
     "apps/api/src/agent-surface/creator-agent-key-routes.test.ts": 635,
     "apps/api/src/agent-surface/creator-agent-key-routes.ts": 460,
+    "apps/api/src/agent-surface/example-files.test.ts": 116,
+    "apps/api/src/agent-surface/example-files.ts": 349,
     "apps/api/src/agent-surface/kit-files.test.ts": 83,
     "apps/api/src/agent-surface/kit-files.ts": 552,
     "apps/api/src/agent-surface/mcp-debug-log.test.ts": 27,
@@ -69,7 +73,7 @@
     "apps/api/src/catalog/games-repo-contract-check.test.ts": 291,
     "apps/api/src/catalog/games-repo-contract-check.ts": 1172,
     "apps/api/src/catalog/github-client.test.ts": 846,
-    "apps/api/src/catalog/github-client.ts": 4294,
+    "apps/api/src/catalog/github-client.ts": 3970,
     "apps/api/src/catalog/local-games-repo.test.ts": 75,
     "apps/api/src/catalog/local-games-repo.ts": 373,
     "apps/api/src/catalog/published-slugs.test.ts": 31,
@@ -108,10 +112,8 @@
     "apps/api/src/community/suggestions.ts": 1024,
     "apps/api/src/community/votes.test.ts": 50,
     "apps/api/src/community/votes.ts": 175,
-    "apps/api/src/creation/agent-tasks.test.ts": 197,
-    "apps/api/src/creation/agent-tasks.ts": 1135,
     "apps/api/src/creation/builder.test.ts": 27,
-    "apps/api/src/creation/builder.ts": 217,
+    "apps/api/src/creation/builder.ts": 159,
     "apps/api/src/creation/code-lane.test.ts": 376,
     "apps/api/src/creation/code-lane.ts": 2069,
     "apps/api/src/creation/code-surface.ts": 216,
@@ -129,8 +131,6 @@
     "apps/api/src/creation/editor-contract.ts": 759,
     "apps/api/src/creation/editor-drafts.test.ts": 258,
     "apps/api/src/creation/editor-drafts.ts": 1144,
-    "apps/api/src/creation/example-files.test.ts": 116,
-    "apps/api/src/creation/example-files.ts": 349,
     "apps/api/src/creation/game-seed.test.ts": 372,
     "apps/api/src/creation/game-seed.ts": 1587,
     "apps/api/src/creation/job-admin-routes.test.ts": 203,
@@ -138,7 +138,7 @@
     "apps/api/src/creation/job-costs.test.ts": 230,
     "apps/api/src/creation/job-costs.ts": 805,
     "apps/api/src/creation/job-state.test.ts": 487,
-    "apps/api/src/creation/job-state.ts": 2034,
+    "apps/api/src/creation/job-state.ts": 2067,
     "apps/api/src/creation/module-size.ts": 128,
     "apps/api/src/creation/owner-games.ts": 98,
     "apps/api/src/creation/quota-gate.ts": 223,
@@ -156,9 +156,8 @@
     "apps/api/src/creation/seed-bundle.ts": 302,
     "apps/api/src/creation/seed-context.test.ts": 142,
     "apps/api/src/creation/seed-context.ts": 461,
+    "apps/api/src/creation/seed-provider-env.ts": 123,
     "apps/api/src/creation/session-crash.test.ts": 14,
-    "apps/api/src/creation/source-patch.test.ts": 25,
-    "apps/api/src/creation/source-patch.ts": 320,
     "apps/api/src/creation/symbol-map.ts": 578,
     "apps/api/src/creation/type-check.test.ts": 153,
     "apps/api/src/creation/type-check.ts": 606,
@@ -272,6 +271,8 @@
     "apps/api/src/platform/server.ts": 59,
     "apps/api/src/platform/slug.test.ts": 135,
     "apps/api/src/platform/slug.ts": 596,
+    "apps/api/src/platform/source-patch.test.ts": 25,
+    "apps/api/src/platform/source-patch.ts": 320,
     "apps/api/src/platform/spa-paths.test.ts": 101,
     "apps/api/src/platform/spa-paths.ts": 440,
     "apps/api/src/platform/store.test.ts": 394,
