+++ eslint-rules/module-size-baseline.json
@@ -43,18 +43,25 @@
     "apps/api/src/agent-surface/agent-game-key.test.ts": 115,
     "apps/api/src/agent-surface/agent-game-key.ts": 242,
     "apps/api/src/agent-surface/agent-session-revocation.ts": 23,
+    "apps/api/src/agent-surface/agent-tasks.test.ts": 272,
+    "apps/api/src/agent-surface/agent-tasks.ts": 377,
     "apps/api/src/agent-surface/agent-upload-token.test.ts": 87,
     "apps/api/src/agent-surface/agent-upload-token.ts": 164,
+    "apps/api/src/agent-surface/build-prompt.test.ts": 171,
+    "apps/api/src/agent-surface/build-prompt.ts": 186,
     "apps/api/src/agent-surface/creator-agent-key-routes.test.ts": 938,
     "apps/api/src/agent-surface/creator-agent-key-routes.ts": 190,
+    "apps/api/src/agent-surface/example-files.test.ts": 166,
+    "apps/api/src/agent-surface/example-files.ts": 198,
     "apps/api/src/agent-surface/game-manifest-hint.test.ts": 87,
     "apps/api/src/agent-surface/game-manifest-hint.ts": 78,
+    "apps/api/src/agent-surface/gate-verdict.ts": 62,
     "apps/api/src/agent-surface/kit-digest.test.ts": 278,
     "apps/api/src/agent-surface/kit-digest.ts": 406,
     "apps/api/src/agent-surface/kit-files.test.ts": 322,
     "apps/api/src/agent-surface/kit-files.ts": 633,
     "apps/api/src/agent-surface/managed-agent.test.ts": 57,
-    "apps/api/src/agent-surface/managed-agent.ts": 215,
+    "apps/api/src/agent-surface/managed-agent.ts": 218,
     "apps/api/src/agent-surface/managed-availability.test.ts": 123,
     "apps/api/src/agent-surface/managed-availability.ts": 142,
     "apps/api/src/agent-surface/managed-backend.test.ts": 482,
@@ -62,7 +69,7 @@
     "apps/api/src/agent-surface/managed-provider-anthropic.test.ts": 238,
     "apps/api/src/agent-surface/managed-provider-anthropic.ts": 267,
     "apps/api/src/agent-surface/managed-provider-copilot.test.ts": 256,
-    "apps/api/src/agent-surface/managed-provider-copilot.ts": 138,
+    "apps/api/src/agent-surface/managed-provider-copilot.ts": 143,
     "apps/api/src/agent-surface/managed-provider-gemini.test.ts": 234,
     "apps/api/src/agent-surface/managed-provider-gemini.ts": 220,
     "apps/api/src/agent-surface/managed-provider-openai.test.ts": 238,
@@ -78,7 +85,7 @@
     "apps/api/src/agent-surface/mcp-server-discovery.test.ts": 132,
     "apps/api/src/agent-surface/mcp-server-discovery.ts": 96,
     "apps/api/src/agent-surface/mcp-server.test.ts": 3674,
-    "apps/api/src/agent-surface/mcp-server.ts": 1722,
+    "apps/api/src/agent-surface/mcp-server.ts": 1728,
     "apps/api/src/agent-surface/mcp-session-key.test.ts": 69,
     "apps/api/src/agent-surface/mcp-session-key.ts": 173,
     "apps/api/src/agent-surface/mcp-session-nudges.test.ts": 177,
@@ -167,9 +174,6 @@
     "apps/api/src/community/votes.test.ts": 284,
     "apps/api/src/community/votes.ts": 92,
     "apps/api/src/copilot-mcp-deploy.test.ts": 40,
-    "apps/api/src/creation/agent-state.ts": 20,
-    "apps/api/src/creation/agent-tasks.test.ts": 272,
-    "apps/api/src/creation/agent-tasks.ts": 377,
     "apps/api/src/creation/builder.test.ts": 178,
     "apps/api/src/creation/builder.ts": 119,
     "apps/api/src/creation/chat-agent.test.ts": 351,
@@ -195,8 +199,6 @@
     "apps/api/src/creation/editor-contract.ts": 1392,
     "apps/api/src/creation/editor-drafts.test.ts": 699,
     "apps/api/src/creation/editor-drafts.ts": 646,
-    "apps/api/src/creation/example-files.test.ts": 166,
-    "apps/api/src/creation/example-files.ts": 198,
     "apps/api/src/creation/game-seed.test.ts": 969,
     "apps/api/src/creation/game-seed.ts": 761,
     "apps/api/src/creation/job-admin-routes.test.ts": 430,
@@ -204,7 +206,7 @@
     "apps/api/src/creation/job-costs.test.ts": 188,
     "apps/api/src/creation/job-costs.ts": 215,
     "apps/api/src/creation/job-state.test.ts": 428,
-    "apps/api/src/creation/job-state.ts": 517,
+    "apps/api/src/creation/job-state.ts": 559,
     "apps/api/src/creation/knowledge-search.test.ts": 491,
     "apps/api/src/creation/knowledge-search.ts": 533,
     "apps/api/src/creation/module-size.test.ts": 70,
@@ -241,8 +243,6 @@
     "apps/api/src/creation/seed-stream.ts": 35,
     "apps/api/src/creation/session-crash.test.ts": 76,
     "apps/api/src/creation/session-crash.ts": 41,
-    "apps/api/src/creation/source-patch.test.ts": 329,
-    "apps/api/src/creation/source-patch.ts": 313,
     "apps/api/src/creation/symbol-map.test.ts": 77,
     "apps/api/src/creation/symbol-map.ts": 257,
     "apps/api/src/creation/tab-complete.test.ts": 96,
@@ -252,8 +252,6 @@
     "apps/api/src/creation/typecheck-preflight.test.ts": 199,
     "apps/api/src/creation/typecheck-preflight.ts": 246,
     "apps/api/src/creator-studio-workspace.test.ts": 321,
-    "apps/api/src/delivery/build-prompt.test.ts": 171,
-    "apps/api/src/delivery/build-prompt.ts": 186,
     "apps/api/src/delivery/build-transcript.test.ts": 339,
     "apps/api/src/delivery/build-transcript.ts": 162,
     "apps/api/src/delivery/games-store.test.ts": 1128,
@@ -268,14 +266,13 @@
     "apps/api/src/delivery/gate-screenshot.ts": 52,
     "apps/api/src/delivery/gate-trigger.test.ts": 201,
     "apps/api/src/delivery/gate-trigger.ts": 281,
-    "apps/api/src/delivery/gate-verdict.ts": 62,
     "apps/api/src/delivery/gcs-sign.test.ts": 62,
     "apps/api/src/delivery/gcs-sign.ts": 193,
     "apps/api/src/delivery/recent-builds.ts": 50,
     "apps/api/src/delivery/source-delivery.test.ts": 606,
-    "apps/api/src/delivery/source-delivery.ts": 540,
+    "apps/api/src/delivery/source-delivery.ts": 545,
     "apps/api/src/delivery/stage-hints.test.ts": 334,
-    "apps/api/src/delivery/stage-hints.ts": 186,
+    "apps/api/src/delivery/stage-hints.ts": 192,
     "apps/api/src/delivery/staged-preview.test.ts": 692,
     "apps/api/src/delivery/staged-preview.ts": 614,
     "apps/api/src/erase-verification-workflow.test.ts": 85,
@@ -323,10 +320,11 @@
     "apps/api/src/platform/admin-session.ts": 10,
     "apps/api/src/platform/admin.test.ts": 1215,
     "apps/api/src/platform/admin.ts": 979,
+    "apps/api/src/platform/agent-state.ts": 20,
     "apps/api/src/platform/agent-token.test.ts": 117,
     "apps/api/src/platform/agent-token.ts": 313,
     "apps/api/src/platform/app.test.ts": 344,
-    "apps/api/src/platform/app.ts": 1143,
+    "apps/api/src/platform/app.ts": 1166,
     "apps/api/src/platform/apple-account.test.ts": 137,
     "apps/api/src/platform/apple-account.ts": 81,
     "apps/api/src/platform/apple-auth.test.ts": 203,
@@ -394,6 +392,8 @@
     "apps/api/src/platform/slug-ownership.ts": 48,
     "apps/api/src/platform/slug.test.ts": 65,
     "apps/api/src/platform/slug.ts": 133,
+    "apps/api/src/platform/source-patch.test.ts": 329,
+    "apps/api/src/platform/source-patch.ts": 316,
     "apps/api/src/platform/spa-paths.test.ts": 121,
     "apps/api/src/platform/spa-paths.ts": 137,
     "apps/api/src/platform/store.test.ts": 799,
